backend/pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "learnops-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "lerkush",email = "lutfiyye@gmail.com"}
]
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.132.0,<0.133.0)",
    "uvicorn[standard] (>=0.41.0,<0.42.0)",
    "sqlalchemy[asyncio] (>=2.0.46,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "cryptography (>=46.0.5,<47.0.0)",
    "itsdangerous (>=2.2.0,<3.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "httpx (>=0.28.1,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "ruff (>=0.15.2,<0.16.0)",
    "mypy (>=1.19.1,<2.0.0)"
]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.mypy]
python_version = "3.13"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
